missed files added

+++ backend/src/Store.Core/ViewModel/EstimationLogsModel.cs
@@ -3,10 +3,10 @@
     public class EstimationLogsModel: BaseEntity
     {
         public long AppUserId { get; set; }
-        public decimal PricePerGram { get; set; }
-        public decimal Weight { get; set; }
-        public decimal TotalPrice { get; set; }
-        public decimal Discount { get; set; }
+        public double PricePerGram { get; set; }
+        public double Weight { get; set; }
+        public double TotalPrice { get; set; }
+        public double Discount { get; set; }
 
     }
 }
